+++ CHANGELOG.md
@@ -19,6 +19,12 @@ We use the following categories for changes:
 
 - Run timescaledb-tune with the promscale profile [#1601]
 
+## [prom-migrator] Unreleased
+
+### Added
+- Support for passing custom HTTP headers via command line arguments for both
+  reader and writer [#1020].
+
 ## [0.14.0] - 2022-08-30
 
 ### Added